tests done in group_member_junction_dao.py for user list and leaving group as well as test

+++ PythonAPI/data_access_layer/implementation_classes/group_member_junction_dao.py
@@ -5,6 +5,7 @@ from PythonAPI.data_access_layer.abstract_classes.group_member_junction_abs impo
 from PythonAPI.entities import group_member_junction
 from PythonAPI.entities.group_member_junction import GroupMemberJunction
 from PythonAPI.util.database_connection import connection
+from custom_exceptions.group_member_junction_exceptions import *
 
 
 class GroupMemberJunctionDao(GroupMemberJunctionAbs):
@@ -21,16 +22,23 @@ class GroupMemberJunctionDao(GroupMemberJunctionAbs):
             for member in group_record:
                 group_list.append(GroupMemberJunction(*member))
             return group_list
-        except TypeError as e:
-            print("you have input the incorrect amount of arguments")
+        except AssertionError:
+            raise AssertionError("This is not allowed")
+        except TypeError:
+            raise TypeError("you have put the incorrect amount of arguments")
 
     # deletes user from group_member_junction_table
     def leave_group(self, user_id: int, group_id: int):
-        try:
-            sql = "delete from group_member_junction_table where user_id = %s and group_id = %s"
-            cursor = connection.cursor()
-            cursor.execute(sql, [user_id, group_id])
-            connection.commit()
-            return True
-        except TypeError as e:
-            print("please check your arguments")
+        check_list = self.get_all_users_in_a_group()
+        for checks in check_list:
+            if checks.group_id == group_id:
+                if checks.user_id == user_id:
+                    try:
+                        sql = "delete from group_member_junction_table where user_id = %s and group_id = %s"
+                        cursor = connection.cursor()
+                        cursor.execute(sql, [user_id, group_id])
+                        connection.commit()
+                        return True
+                    except TypeError:
+                        raise TypeError("too many arguments")
+        raise WrongId("Incorrect ID")
